+++ sdk/data/cosmos/azcosmos_emulator_tests/emulator_cosmos_database_test.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/cosmos/azcosmos"
+	"github.com/Azure/azure-sdk-for-go/sdk/data/cosmos/azcosmos"
 )
 
 func TestDatabaseCRUD(t *testing.T) {
