+++ sdk/data/cosmos/azcosmos_emulator_tests/go.mod
@@ -1,7 +1,7 @@
-module github.com/Azure/azure-sdk-for-go/sdk/cosmos/azcosmos_emulator_tests
+module github.com/Azure/azure-sdk-for-go/sdk/data/cosmos/azcosmos_emulator_tests
 
 go 1.13
 
-replace github.com/Azure/azure-sdk-for-go/sdk/cosmos/azcosmos => ../azcosmos
+replace github.com/Azure/azure-sdk-for-go/sdk/data/cosmos/azcosmos => ../azcosmos
 
-require github.com/Azure/azure-sdk-for-go/sdk/cosmos/azcosmos v1.0.0
+require github.com/Azure/azure-sdk-for-go/sdk/data/cosmos/azcosmos v1.0.0
